\name{biovar-package}
\alias{biovar-package}
\alias{biovar}
\docType{package}
\title{
Estimation of biological variation
}
\description{
Estimates biological variation components utilizing two-fold nested ANOVA models, Bayesian modelling and Bootstrapping.
}
\details{
In order to streamline biological variation estimation based on BV data, this package will provide all necessary modern methods to do so.
}
\author{
Author: Pernille Kjeilen Fauskanger
Maintainer: Pernille Kjeilen Fauskanger
}
\references{
Books
}
\keyword{ package }
\seealso{
Look at the horse. It does the work. Wow.
}
\examples{
# Here you can print stuff
print(1)
}
