% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_stan_output.R
\name{calculate_cvs_identity_scale}
\alias{calculate_cvs_identity_scale}
\title{Calculate CVs assuming a model on the original (identity) scale.}
\usage{
calculate_cvs_identity_scale(fit)
}
\arguments{
\item{fit}{A list of posterior draws.}
}
\value{
A list containing posterior draws for each calculated CV.
}
\description{
Calculate CVs assuming a model on the original (identity) scale.
}
