% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_stan_output.R
\name{calculate_cvs_log_scale}
\alias{calculate_cvs_log_scale}
\title{Calculate CVs assuming a log-transformed model.}
\usage{
calculate_cvs_log_scale(fit)
}
\arguments{
\item{fit}{A list of posterior draws.}
}
\value{
A list containing posterior draws for each calculated CV.
}
\description{
Calculate CVs assuming a log-transformed model.
}
