% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bootstrap_output.R
\name{process_bootstrap_output}
\alias{process_bootstrap_output}
\title{Process and Summarize Bootstrap Model Output}
\usage{
process_bootstrap_output(
  fit,
  log_transformed,
  analyte = "X",
  group = "Y",
  data = NULL,
  digits = 1L
)
}
\arguments{
\item{fit}{An \code{list}. The output from \code{bv_anova_bootstrap_ci()}}

\item{log_transformed}{A \code{logical} value. If \code{TRUE}, it assumes
the model was fitted on log-transformed data, and results
(like \code{beta}) are back-transformed. CV calculations are also adjusted
accordingly.}

\item{analyte}{A \code{character} string used to label the measurand in the
output tables. Defaults to "X".}

\item{group}{A \code{character} string used to label the group or cohort in
the output tables. Defaults to "Y".}

\item{data}{An optional \code{data.frame} or \code{data.table} containing the
original input data. If provided, its \code{SubjectID} column will be used
to label the subjects in the subject-specific output table.
Defaults to \code{NULL}.}

\item{digits}{An \code{integer} specifying the number of decimal places to use
when formatting the confidence intervals. Defaults to 1.}
}
\value{
A \code{data.table}.
}
\description{
Process and Summarize Bootstrap Model Output
}
\examples{
print(1)
}
