% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_bv_data}
\alias{simulate_bv_data}
\title{Simulate Biological Variation Data}
\arguments{
\item{n}{\code{Integer}. The number of subjects.}

\item{S}{\code{Integer}. The number of samples taken from each subject. Defaults to 10.}

\item{R}{\code{Integer}. The number of replicates measured for each sample. Defaults to 2.}

\item{cvi}{\code{Double}. The coefficient of variation (in percent) for variability within each subject (default is 10).}

\item{cva}{\code{Double}. The coefficient of variation (in percent) for the variability of analytical error (default is 2). This is the source of variation between replicated measurements.}

\item{cvg}{\code{Double}. The coefficient of variation (in percent) for variability between subjects (default is 50).}

\item{mu}{\code{Double}. The overall true mean for the population (default is 100).}
}
\value{
A `list` with columns:
\itemize{
  \item \code{SubjectID} - The ID of the subject.
  \item \code{SampleID} - The ID of the sample.
  \item \code{ReplicateID} - The ID of the replicate.
  \item \code{y} - The measured value for the replicate.
}
}
\description{
This function simulates biological variation data based on a specified number of subjects, samples, and replicates.
}
\details{
Simulate Biological Variation Data

This function generates simulated biological variation data. It first generates subject-specific means based on a normal distribution with mean `mu` and coefficient of variation `cvg`. For each subject, it then generates `S` samples based on their specific mean and the coefficient of variation `cvi`. Finally, for each sample, it generates `R` replicate measurements based on the sample mean and the coefficient of variation `cva`.
}
\examples{
\dontrun{
# Simulate data for 100 subjects, with 10 samples per subject and 3 replicates per sample
output <- simulate_bv_data(15, 10, 2, 10, 2, 50, 100)
print(output)
}

}
