\name{biovar-package}
\alias{biovar-package}
\alias{biovar}
\docType{package}
\title{
\packageTitle{biovar}
}
\description{
\packageDescription{biovar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{biovar}
\packageIndices{biovar}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{biovar}

Maintainer: \packageMaintainer{biovar}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
