\name{biovar-package}
\alias{biovar-package}
\alias{biovar}
\docType{package}
\title{
\packageTitle{biovar}
}
\description{
\packageDescription{biovar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{biovar}
\packageIndices{biovar}
No
}
\author{
\packageAuthor{biovar}

Maintainer: \packageMaintainer{biovar}
}
\references{
Books
}
\keyword{ package }
\seealso{
Look at the horse. It does the work. Wow.
}
\examples{
# Here you can print stuff
print(1)
}
